requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
pydantic==2.5.0
requests==2.31.0
httpx==0.25.2

# Document processing
PyMuPDF==1.23.8
python-docx==1.1.0
pypdf2==3.0.1

# Vector embeddings and search
sentence-transformers==2.7.0
huggingface_hub==0.20.3
transformers==4.36.2
torch==2.1.2
faiss-cpu==1.7.4
numpy==1.24.3
scikit-learn==1.3.2

# LLM integration
google-generativeai==0.7.2

# Text processing
nltk==3.8.1
spacy==3.7.2
beautifulsoup4==4.12.2
lxml==4.9.3

# Database (optional for caching)
sqlalchemy==2.0.23

# Utilities
python-dotenv==1.0.0
pydantic-settings==2.1.0
loguru==0.7.2
tqdm==4.66.1
aiofiles==23.2.0
